Code cleanup + Fixes

+++ src/main/java/com/zpedroo/voltzevents/utils/scoreboard/manager/ScoreboardManager.java
@@ -44,6 +44,8 @@ public class ScoreboardManager {
     }
 
     public static void removeScoreboard(Player player) {
+        if (player == null) return;
+
         ScoreboardHandler scoreboard = scoreboards.remove(player);
         if (scoreboard != null) {
             scoreboard.delete();
